# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5Swagger",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local"
        },
        {
            "url": "http://localhost/api/auth",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/bill": {
            "get": {
                "tags": [
                    "Счета"
                ],
                "summary": "Получить все счета",
                "description": "Получить счета",
                "operationId": "164f5f130f27a6f488d4a56c9310a5d6",
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Bill"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bill/{id}": {
            "get": {
                "tags": [
                    "Счета"
                ],
                "summary": "Получить счет",
                "description": "Получает счет по идентификатору и возвращает его",
                "operationId": "3e993d95894c63cc6a243070ed1b4f85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор счета",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ff7f36b1-1cab-35b9-9b3f-969bb0e92109"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Bill"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Счета"
                ],
                "summary": "Обновить статус счета",
                "description": "Обновляет статус счета",
                "operationId": "ff35dd0923e3f82b925ac6d648e1d397",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор счета",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ff7f36b1-1cab-35b9-9b3f-969bb0e92109"
                        }
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Новый статус счета",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "closed"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/BillStatus"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Bill"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bill/create": {
            "post": {
                "tags": [
                    "Счета"
                ],
                "summary": "Создать счет",
                "description": "Создает новый счет и возвращает его",
                "operationId": "b2c9f9e0eb3497aec544bc0f1bd4178e",
                "parameters": [
                    {
                        "name": "arendators_count",
                        "in": "path",
                        "description": "Количество пользователей связанных со счётом",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "balance",
                        "in": "path",
                        "description": "Баланс счёта",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "48658.52"
                        }
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Тип счёта",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "personal"
                        }
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Статус счёта",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "open"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/BillCreate"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Bill"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/bill/{id}/delete": {
            "delete": {
                "tags": [
                    "Счета"
                ],
                "summary": "Удалить счет",
                "description": "Удаляет запись о счете",
                "operationId": "4e2aec65efff56c64e418afa29fe12c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор счета",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ff7f36b1-1cab-35b9-9b3f-969bb0e92109"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Bill"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Bill": {
                "title": "Bill",
                "properties": {
                    "id": {
                        "type": "uuid",
                        "example": "ff7f36b1-1cab-35b9-9b3f-969bb0e92109"
                    },
                    "arendators_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "balance": {
                        "type": "decimal",
                        "example": 48658.52
                    },
                    "type": {
                        "type": "string",
                        "example": "personal"
                    },
                    "status": {
                        "type": "string",
                        "example": "open"
                    }
                },
                "type": "object"
            },
            "BillCreate": {
                "title": "BillCreate",
                "properties": {
                    "arendators_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "balance": {
                        "type": "decimal",
                        "example": 48658.52
                    },
                    "type": {
                        "type": "string",
                        "example": "personal"
                    },
                    "status": {
                        "type": "string",
                        "example": "open"
                    }
                },
                "type": "object"
            },
            "BillStatus": {
                "title": "BillStatus",
                "properties": {
                    "id": {
                        "type": "uuid",
                        "example": "ff7f36b1-1cab-35b9-9b3f-969bb0e92109"
                    },
                    "status": {
                        "type": "string",
                        "example": "closed"
                    }
                },
                "type": "object"
            },
...
            "Response401": {
                "title": "Response401",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
            "Response403": {
                "title": "Response403",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden"
                    }
                },
                "type": "object"
            },
            "Response404": {
                "title": "Response404",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}